Add variable parsing and line-number tracking
This change started by adding `var` assignment tracking.  Local assignments tokenize and compile (untested) correctly, but do not yet execute correctly.

As part of looking to debug a longer file, I then went to include line number information on Token.  That required wrapping Token in a new struct ParseToken to carry the line number and source offset information, which is also currently untested/unused.

var, = and == are now tokenized (and partially used).

ParseToken now wraps Token and holds the byte_range to be used to get back into the source to grab the token name as well as the line number.  This allows deleting Token.name() which was a hack anyway.

InputManager now carries the line_number and current token offset.  Tokens are now made using InputManager.make_token rather than just Token::Foo() directly.  The Token enum is now a combination of Token Type and Value in wren_c parlance.

InputManger.next updates line_number on every \n.  (Untested).

Compiler now tracks locals as they are seen.  Not clear that the emitted bytecode is correct however.  In wren_c, locals just end up on the stack so there is no locals information passed between the compiler and the vm, just rather the emitted bytecodes are built with the right lookups relative to the function-local top of stack.  None of this exists for wren_rust yet.

Pass can_assign through all PrefixParslets.  Only used for name/bare_name, and unclear if correctly yet.