Fix compile error output to match test.py expectations
This "passes" 46 more tests, even though many of the passes are erroneous.

The harness checks for compile errors (vs. runtime errors) and line numbers, but it does not distinguish types of compile errors (e.g. lexer vs. grammar) so many of these "passes" are just lexer errors (we don't yet know how to handle the given token) but the harness is treating them as "passes" for these error tests, now that our harness is printing out errors in the expected format.