Add support for parsing minus
Started by just trying to add support for minus (-) and then blossumed into making number parsing more real and also fixing numbers to be stored as f64 (doubles) rather than u64 (unsigned integers) as they have been until now.

As part of adding minus parsing, also added method name symbol registration, which is global on the vm object, so required exposing the vm object in the parser.  Since I can't borrow both the vm and its module(s) separately, Parser now only exposes the vm object and not a direct reference to the module.

Experimented with improved error handling.  Some is better, some is worse.  The better part is the implicit conversion from std error types to LexError, that's very clean.  The worse part is that now we have 4 separate enums on LexError to wrap these which may be unecessary.

I also don't know how to trigger the auto-wrapping cleanly, outside of ?, so ParseToken.name is a bit uglier.

Added Input.peek_* variants which take the thing to check against (either u8 or closure) and made implementing read_number while handling newlines much easier.

Similarly added skip_while on InputManager.  It's slowly becoming an Iterator.  Perhaps it just becomes the TokenIterator, or Tokenizer. :)

Taugh the lex() helper how to return a WrenError.  I realized that if InputManager becomes an iterator, then this just becomes a collect() call over the InputManager.

Since - is implemented as a unary op, I had to add support for registering its method name, which required the above VM reference changes.

It's not clear if Signatures will be needed in my version?   In wren_c there is a methodname table which maps from name to integer for passing.  Unclear where they're read.

Previous lexer or grammar errors around '-' are now MethodNotFound as it's attempting to call the '-' method but I haven't added a primitive for it.  That should be an easy next patch.