Add first-pass support for "class" keyword
Passes 11 more tests!

This was a whole can of worms.  Required implementing superclass and metaclass class hierarchy (which is good, we can now wire up static methods on top of this).  To wire up that required splitting out classes off of the WrenVM object and onto a CoreClasses object which could be built separately after the rest of the VM is initialized.

I transliterated parts of the class logic from wren_c, but a lot of it doesn't make sense (yet?) in wren_rust.  My guess is we will end up with something much simpler in rust.

There are a whole bunch of failures now do to missing newlines, which is probably an error in the class grammar.

Class uses a Placeholder op, just like jumps do and later gets filled in with the number of fields a class has.  Since we don't yet support fields, we just always send 0 (but we do go through the logic of replacing the opcode).

Made Compiler.emit() a bit more generic, including returning the index of the added op.  This made it possible to remove the two other emit_*_placeholder calls.

Added consume_at_least_one_line (win wren_c it's confusingly called consume_line) and replaced the one place in blocks it was expected but not used.

Taught add_local and declare_named_variable to return the "symbol" / index used (matches wren_c).

define_variable is stubbed out, but not sure that it makes sense in the end.  It really only handles the module case (even in wren_c) unlike what it name suggests.

Added basic class_definition(), could be simpler in rust I think.  Still needs refactoring.  Only supports parsing the class block, doesn't actually do anyting with the things inside it.

(Temporarily) no longer using the module_name passed into compile().  It always dummy_module anyway and with the current location of the core parsing (mostly due to having only a single module ever) it was causing a reset of the module, losing all the loaded core classes.

Moved away from Value::Object for now. Split to Value::Class and Value::Range (and will likely have a few other types).  Not sure what this means once we start instantiating user objects, but for now this works.

Attempted to move more visibility to pub(crate) instead of pub.  Probably too early to try this, but I'm attempting to get more in the habit.  Eventually will want to create an explicit public API.

Added Value.try_into_string() for reading the string out of the Value pushed on the stack from the class token.

Split class definitions out into CoreClasses so it can be set later on WrenVM after the rest is initialized.

Have a basic wren_new_class. Probably can be very short once I'm confident I don't need much of the wren_c bits.

Added as_class which returns the Handle<ObjClass> from a Value or panics.  Not sure if this is better/worse than the try_into_* which are used in other places.

Added Handle<T> as an alias for Rc<RefCell<T>> and cleaned up the code a ton.  Thank you @abarth for the suggestion.

Added wren_define_variable, but did not attempt the implicit variable case yet. Not sure if other code even creates implicit variables yet.

Added init_core_classes which wires up Object, Class, etc. from wren_core.wren.  This is currently called during WrenVM init, but may be moved later.  It also currently loads into the one module, instead of having a separate module for core.