Add first-pass support for "for in"
This doesn't fully work.  It's not handling the local iterator value correctly (so the test added just prints 0, 0, 0, rather than 0, 1, 2).  But it loops the right number of times at least!

Also added Range.iterate(_) and Range.iteratorValue(_)

Fixed a bug in the iterate debug code where it would try to print a module offset as a local offset and reach for an index outside the stack bounds.

I started to add more handling for scopes but then commented the code out to be able to land this.