Fix errors to report the line number of the last token parsed
Previously it was using the last token created (which is 2 tokens ahead of now), instead of the last token looked at.  This fixes a bunch of tests which were previously only wrong in their error line number.

Still a lot of work to do on making error messages match wren_c.