Add basic stub_wren_core.wren as part of initialization
Added sub_wren_core.wren and wired it into vm initialization.

Since we only have a single module and single fiber (thus single stack and pc) for now, there were some hacks involved in wiring. :)  Including the fact that I didn't try to write a script to convert the .wren into a string to compile it into the binary.

Also fixed class and object_metaclass to be correctly wired up to their superclasses (e.g. inherit methods).  This was the cause of two panics in the tests which are now resolved.

137 tests passing!