Refactor to passing ParseContext around instead of Parser
Parser having a mutable reference to VM was awkward (but necessary in order to access the module).   It also prevented us from being able to swap out what compiler was passed around.

This refactor should theoretically make it possible to create a new ParseContext (or update the current) replacing the Compiler reference with a new one.  Will test in the next patch.

Had to move all methods out of Parser (which matches wren_c) to allow access to the vm which is now on the ParserContext rather than the Parser.  The VM is needed for the module name for the parse_error method.