Add function parsing
This was a really hard patch.  As such, it came in with a bunch of extra junk as I mostly just need to land.

I spent a lot of time fighting tracking Compilers (in a list/stack).

I also found a bug in wren_c (I think?) wereby wren_c has a null class pointer for functions/closures defined inside wren_core.wren.  I'm not sure if this is observable or not since I don't think wren has tearoffs through which to grab the type of wren_core.wren defined functions/methods.

Fn is now declared (twice actually, that's a bug), which fixes a bunch more tests.  Block arguments are kinda supported?  (enough to not crash).  compiler.rs now knows how to have more than one Compiler and auto-pop the compiler on errors too.

There is now a Closure Op, but it doesn't actually do anything.  We don't yet support defining or executing block methods.  Closure will just throw null on the stack so that the vm won't crash due to insufficient stack.

Some Upvalue code was written (and also often commented out).  They're not ready for full implementation, but there are some pieces in place.

ParseContext now owns the Compiler "stack" (really a list).  have_compiler is a bit of a hack, because at the end of an exhausting change I couldn't quite get the borrows or matches right to achieve the same in the one place it's used.

Added ObjFn so that end_compiler could return it.  I have not yet added ObjClosure yet.

Nothing actually ever uses/executes the compiled function code.

PushCompiler uses a Drop trait (thanks @abarth) to handle auto-poping the compiler when errors happen.  the normal path is to have an explicit pop where the compiler is then passed to end_compiler.

end_compiler is a bit sketchy in that it assumes some non-local state, such that ctx already points to the parent compiler.

method_call now has some basic support for parsing block arguments.

Commented out ObjType for now.  Probably can be removed.