Partial implementation of Fn.call(*)
This adds Fn.call for arity 0-16.

I don't think function calls are actually working though?  As I would expect more tests to pass.  But this is a commit-able start.

This changes WrenVM.run_fiber to be run_frame_in_fiber, and pass the frame.  The Fiber is currently not even used.  Calling a function now returns out to run() which catches the RunNext::FunctionCall return type and does the frame manipulation before calling back to run_frame_in_fiber in a loop.  wren_c just has a big loop in one function and modifies the frame variables on the stack.   This is  a bit more awkward, but "safer" and more rusty, I think?

It's unclear if RunNext::Done is ever a valid return type, or if that should panick.

CallFrame now holds the stack, pc and closure being executed.  Fiber holds the stack of CallFrames (for some later ability to interrupt/resume fibers?) and WrenVM just holds the executing Fiber.

WrenVM now has a last_fiber, but it's not clear it's ever useful (since I doubt it has anything on it when it would be used, the CallFrame is probably gone by then?)  I'll probably remove it.