Make basic function execution work!
Previously we were never pushing the calling frame onto the call_frame stack, so we would just get to the end of the called function and end. :)

Additionally once I fixed that, we were tearing off the pc from the frame and never restoring it, so we looped infinitely once it was possible to return to the calling frame.  Fixed for now by just modifying the pc on the frame instead of tearing it off.

3 more tests pass!