Add ObjInstance (now possible to instantiate objects)
Woohoo!  6 more tests pass and a whole new world of possibilities.

Fields don't yet work, but ObjInstance lives.  Also fixed constructor methods when run on the instance to return this instead of null.

It's possible this may still be leaking a copy of the instance on the stack?  Not sure.  The arity of the initializer is currently 0, but that's probably wrong?