Support parsing setters
Our implementation of call() was using method_call instead of named_call.  Fixed.  Also fixed ScopePusher to push and pop the class_info, since previously the class_info was just "leaking", once we ever saw a class we would think we were inside a class for the rest of the file.  That will break method parsing in the next change.

Added a bunch of (probably working, but) commented out code about method parsing which I can turn on separately to better see the test changes.

This is all in service of attempting to get wren_core.wren to parse, but that requires being able to call methods which have not yet be declared!