Treat first-char lowercase bare names as method calls
This is just an odd part of wren.  It will treat lowercase bare names as method calls, even if they're not yet declared.  This is required to (eventually) make wren_core.wren parse.  It also changes a whole bunch of "undeclared variable" errors into "could not resolve this" errors.

"could not resolve this" is caused by the compiler not implicitly pushing this as a local.  That code is currently commented out as it was causing other crashes earlier, but can be fixed in the next patch.