Add partially working fields support
I think field identifiers are not compiled correctly hence so many "field out of bounds" errors.

But this does seem to make fields work a little!  And passes one more test (possibly erroneously since it's an error test).

Add Ops for StaticField and Field.  Similarly Tokens for Field and StaticField (since they have different parsing rules).  Added prefix grammar rules for Field, static_field is commented out for now.

Fixed tokenizer to tread names with leading _ or __ as fields.   Refactored read_name to return a String instead of a Token to do this.

Had to make enclosing_class a RefCell to be able to walk the compiler chain with immutable borrows and then also grab a mutable ref to the enclosing_class data while also still holding a ref to the compiler chain (and thus not being able to grab one mutably).  I certainly welcome suggestions as to how to improve this design.

Changed ensure_method to ensure_symbol now that SymbolTable is used for more than just methods.

Added field() grammar.

Fixed more uses of ensure_symbol to use "symbol" as the local variable name.

Added try_to_instance for grabbing the ObjInstance out of a Value.

Uncommented num_fields code now that we're actually needing to lookup fields on ObjInstances (although I'm not sure we're setting them correctly yet).