Rewrite RuntimeError handline to uses Strings instead of enum
wren_c uses strings for errors.  We had used an enum (thinking that was rusty), but it's both become unstainably verbose as well as incompatible with wren_c.

This also adds stack trace generation (not actually sure if it works yet) and brings our wren_test printing closer to what test.py expects.

We still need to implement fully pc -> line number tracking in order to print the correct line numbers for failures (which test.py does test for).  We also currently have a hack in place re: module paths (which test.py cares look like paths, but doesn't care about the contents).

Also moved to using a typedef alias for Result<T> to make core.rs and vm.rs much cleaner.  Moved to using more validate_ functions instead of try_into directly.  Also added this_as_* functions matching wren_c's AS_* macros.  If nothing else we share error messages, even if our this_as_* functions do more work.

There is a lot of (eager) string mallocs in this change, which probably will need to get refactored away.  We don't need to be allocating the strings as early as we do, but it was the easy thing to do for this change and we can remove it once we start benchmarking.

Also changed to_fix.py to show more output types, now that we're passing a lot more tests.

check_arity is implemented (mostly) but commented out, as I tried to minimize the type of change to just the error message replacement without adding new checks in this change.

57 tests are now closer to passing and should pass as soon as we plumb through pc -> line number tracking.