Fix FiberAbort to just print String values
This fixes 1 test.  12 more tests will be fixed as soon as we ignore wren_core.wren from callstack printing.

It's not clear what we're supposed to print for a FiberAbort with a non-string value.  But presumably we'll figure that out when we get to dealing with Fibers.

For now Fiber.abort is just the way that wren_core.wren knows how to report errors.