Make it possible to have multiple modules
This was all in service of plumbing the real module name all the way through.

As part of doing so, I fixed us to have multiple modules.

As an accident to fixing that, the core module name is now separate, meaning ObjFn's compiled into wren_core.wren show up in stack traces with a separate module name, which happens to be #core, which happens to not match the test.py regexp for stacktrace output.... Hence all the tests which were previously "failing" for having found the wrong "line number from first stacktrace line" are now "passing" since the #core lines don't match as stacktraces. :)

The ownership (explained in a comment) is such that the VM owns all Modules.  They're pulled out for compiling and for running and then put back in.  I'm not sure if this will turn out to be correct, but it' compiles!  And runs!  And is probably sufficient for now.

There was a bit of dancing to get core compiled.  We make a stub module, add the basics (Object, Class, etc.) to that, then set that as WrenVM.core_module, actually compile/run stub_core_wren.wren, and then replace the previous core_module with the result of that compile.