Add List.[_]=(_) and fix subscript setter name generation
wren_c has some funny +1/-1 adjustments to arity for subscript setters which I didn't originally understand.  I think what's happening is that arity actually just applies to within the [_] block for setters, and the thing-being-set is always 1, so it needs to both add one for the thing being set and then subtract it again when generating the args underbar/comma string.

Regardless, this now works and passes 9 more tests!