Convert Fiber to ObjFiber and implement Fiber.new()
Fiber.new() actually works now, just nothing to do with it since none of the other Fiber methods are implemented.

This makes 2 more tests pass and all the fiber tests obviously get a lot further. :)

Had to move when Fiber's class was initialized to make it possible to create a Fiber for parsing wren_core.wren w/o it having a null class (which appears to be the route wren_c takes).

I'm sure that this will lead to ref cycles now that instances can get a reference back to the Fiber (or will soon be able to once I implement Fiber.current).