Begin to add a public API (and start to move main.rs to use it)
Begin to add a public API in rust, matching the one for wren_c.  The eventual goal would be to have a public API and move main.rs to a separate crate.  Also probably to expose said public rust API as a public C API matching the wren_c one so that wren_rust could be a drop-in replacement for (and possibly even use directly code like the random or meta optional modules from) wren_c.  However, that's a long ways off, this is a start.  Lets us also start to add module loading logic for wren_test in main.rs in the next few patches.