Move main.rs (wren_test) to using (mostly) public API
This meant moving all of the --tokenize, --compile and --interpret logic out into a new wren_debug.rs module (which will eventually become its own binary), as well as adding a wren_interpret API to wren.rs as well as adding the error_fn accessor on WrenVM.

This appears to pass tests exactly as the previous version did, so that's a success. :)

This should unlock further refactoring to make the wren_test binary its own target as well as wren_rust as library.

Also added comments to the top of all rust files explaining what wren_c equivalent they are based on.

We continue to *intentionally* match the wren_c design/API to not rule out a future possibility of being a drop-in replacement for wren_c embedders.  That also may allow us to share code (e.g. the random or meta modules which only use public API) with wren_c.