Fix Value::hash to not hash all Value::Num to the same value
It turns out our Hash implementation for Value wash hashing the bits of the *reference* to an f64 rather than the f64 itself.  This resulted in unpredictable behavior, as it matter where the Value itself was laid out in memory.

This I think was the reason why churn.wren would sometimes fail in different ways.  It also was causing my attempts to add a HashMap in front of the constants list to take forever.

This change also contains (commented out), all the code needed to enforce the MAX_CONSTANTS limit as well as re-use constant ids (instead of declaring a new constants every time).  I'll land those in a separate change.

Supporting limits and the paired hash/list lookup/recordkeeping caused me to abstract constants behind a ConstantsBuilder object (maybe ends up as part of a FnBuilder?)

ensure_constant is now a free function which takes a ParseContext (which it needs to be able to raise errors).

end_compiler (which probably gets refactored) now calls can throw errors (once MAX_CONSTANT enforcement is turned on), so it also needs to return a Result.