Add basic Fiber.call()
This required making it possible to return a FiberAction from a Primitive, not just a Value.  I made a separate kind of Primitive to make this possible as well as a new macro which knew how to create FiberActionPrimitives rather than ValuePrimitives.

I put call_stack into a RefCell, to allow borrowing it mutably separate from the Fiber itself which otherwise does not need to be borrowed mutably.  Previous attempts at this with a mutably borrowed ObjFiber, would crash, as other access to the ObjFiber (e.g. for its class) would panic while the ObjFiber was mutably borrowed for the duration of the run_in_fiber (only for its call_stack).

validate_fiber_action is the wren_rust equivalent of runFiber.  Instead of running directly, we validate the parameters from the callsite, and then return out to the calling run_fiber function to actually modify the fiber setup.

Moved RuntimeError creation into a RuntimeError::from_vm_error.

FiberRunSource matches wren_c, but is probably the wrong way to store this data. Kept for now.

Added setup_for_call and end_all_take_caller helpers to help setup start/stop call executions of a fiber.

Added fixme about possibly moving away from popping off the top frame during frame execution?  It may no longer be needed if we can borrow the call_stack separately from the ObjFiber.

FiberAction::Call now knows how to start a call.

FiberNext::Return now knows how to end a call (and continue at the caller) or return all the way out when needed.

This feels good enough to commit, but not really settled in design.  We'll see what try, etc. look like.