Add a partially working version of Fiber.suspend
Fiber.suspend is never tested in the wren_c tests.  To test it properly will require working API tests (which we don't quite have until we implement foreign).

I'm checking in this partially working copy with a very basic test (that you can call Fiber.suspend() on the main fiber and it ends the interpreter immediately).

A better test would involve suspending a non-main fiber and then resuming it via the API.