Check for undefined variables (doesn't do anything yet)
Add the plumbing for checking for undefined variables.

The variable definition logic currently errors on undefined instead of setting a fake line number, but this added logic should make that throw the correct errors once I turn it on.

I also changed ParseToken::name to panick instead of Err in the case of an UTF decode error.  As written it should not be possible to ever get an error.

Eventually ParseToken::name should change to not even decode again, but rather just return a slice into the input stream.

The nightly rust build I'm using keeps trying to add imports into std::, so I decided not to fight it this time, so it imported usize into vm.rs.  :shrug: