Uncomment more of stub_wren_core.wren
Now that it's possible to use module variables before they're defined, we should be able to (almost) use an unmodified wren_core.wren.  This is one step in that direction.

Doing this in pieces to try and make it understandable.  Missing String.iterate still blocks a lot of these tests.

Passes 16 more tests!  (No over 600 passing!)