Remove num_traits dependency
This was our only cargo dependency.  I removed it initially for
binary size (no effect), but kept it out for simplicity.
The resulting code is a bit more fragile, but should be fine.

Also removed the debug symbols from release builds.  That moved
the binary size from 8mb to 4mb for wren_test