Fix wren_test and test.py to match wren_c paths
This makes test.py run wren_c tests with wren_c as the cwd for wren_test.

It also changes wren_test to match main.c behavior of adding a leading ./ to paths to make sure they're always expressly relative (instead of bare names which test.c treats differently).

Passes 4 more tests (which we are already passing except for the path in the error message).