Add hacky Map.iterate(_) Map.keyIteratorValue(_) and Map.valueIteratorValue(_)
Map.isEmpty uses Map.iterate(_) which is why this was causing so many tests to fail.

This doesn't work the same way wren_c's does (and doesn't seem to work quite right anyway).  wren_c has a custom hashmap implementation, this just sits on top of rust's.

Map.iterate(_) is clearly not exactly right as iterate.wren does not pass.  Not yet sure why.

Saving this as a checkpoint.  This makes 14 more tests pass.