Fix discard_locals to not modify Compiler.locals
It turns out, discardLocals in wren_c has two uses. One during popScope to issue pops for all going-out-of-scope locals as well as adjust the compiler's locals array, and a second for break/continue where it only issues pops and doesn't adjust the compiler locals array.  I had only implemented the former and when I tried to use it in the latter case it was wrong.

I renamed discard_locals to emit_pops_for_locals to be more clear what it does, and made it not adjust the Compiler.locals array, and instead taught pop_scope how to do that.

This fixed another 5 tests.  Including mandlebrot.wren! which now may be the first "useful" fully working wren program working in wren_rust.  It's definitely slower in wren_rust than wren_c (even in release mode).  Something to fix.

Also changed the way the stack prints during wren_debug interpret mode to be easier to read.  It's closer to the way wren_c prints during their interpreter debugger (although still quite different).

Now that I understand scopes better, it's possible ScopeDepth should be refactored.  The ScopeDepth enum we've chosen is less sentinel-value, but still not quite as clear as it should be.