Add support for unicode escape sequences
wren_c really likes its invalid utf8 testing.  So lots of these tests fail in wren_rust still, since we don't allow invalid utf8 in strings. :/

This seems to add a (possibly working too well) implementation of unicode escape sequences.

Rust does *not* treat "0" as a terminator for strings unlike wren_c, so strings with \0 in them aren't terminated, rather the \0 is printed.

This causes 9 more tests to pass.  And 1 previously "passing" test to fail now that we parse \0 as the byte 0 and then don't use it as a terminator (unlike wren_c).