Fix fiber error handling
This teaches fiber error handling to always set fiber.error when exiting a fiber from an error, as well as walk up the chain of fiber.callers until it either finds a fiber called with try (in which case it returns the error as a string and returns control to that caller), or until it runs out of fibers to return from and returns out of the .run() call all together.

Fixes 3 tests.