Implement Fiber.transferError(_)
This mostly involved tearing the logic from the Err case out into a function so it could be called from transferError as well.

I added a test to cover transferring back after a transferError.  The test document's wren_c's behavior.  This caused me to uncover that transfering back to a fiber which had more than one callframe would fail, so I fixed that too.