Add util/compare.py to easily run wren_rust and wren_c at once
This should help, a lot, in comparing these last few tests.  Eventually maybe I'll event make it diff them.

Also found a bug but have not fixed it yet, will try to see if cargo clippy will find it or if I need to search for others like this manually.

Lots of clippy patches incoming!