Fix Map.iterate(_) to not walk off the end of the map
Map.iterate(_) was going one value too far.

Fixing that caused some of the map tests to "fail" which after further inspection turns out they weren't testing what they were trying to test!

Map.iterate(_) is supposed to return the next value to pass back to it.  Nothing is known about that value.  for wren_c, it returns occupied slots in the hash.  If the first slot in the hash is "5" then 0, 1, 3, 4, will all return "5" from Map.iterate in wren_c.  This was exactly what map/iterate.wren was hitting, as it was checking for one more iterate(_) value than was valid, because assuming 0, 1, 2, 3, 4 are seprate values to map.iterate(_) in wren_c is invalid.  They *are* seprate values to iterate(_) in wren_rust, because we don't return slot ids, we just use rust's own indexing into maps.

So I forked the tests and removed the problematic behavior and skipped the originals.  Eventually this should be upstreamed.