Fix fiber call/try to not push args when not expected.
It turns out the root cause of the return_from_transfer_error
failure was due to the *first* call into the fiber pushing
(and leaking) a Null value on the stack, so later local
lookups failed.

This follows the logic from wren_c, of checking first if the called
fiber expects an arg, and only then pushing the argument value.

I also fixed return_from_transfer_error to have the correct number
of quotes (zero) around the expected string result.