Copy in wren_c's wren_test
This adds a (temporary?) copy of wren_c's wren_test so that I have something I can make build against wren_rust's c API.

Right now this doesn't quite link, but it's also not that far off!
https://gist.github.com/eseidel/91f84e51c4dcf433891ec2a53db01de0 is the burndown list of symbols to add.

I also will probably comment out/remove .c files instead of adding apis on the path to making it link.

The modification to cargo.toml was to tell cargo to build both a normal rust library and a c-style dynamic library (.so).

The premade5.lua added seems to work and generated a valid Makefile.