Add wrenGetListCount
I'm unsure if this should be implemented as part of c_api or in vm.rs itself.  The Rust API eventually will need one of these, but it's unclear if it will want the same assert/panic behavior.  So for now I implemented this in c_api.

Given how much we grab at .elements (on list) or .data (on map) just to check the len, I added len() helpers to both.