Add wrenInitConfiguration and wrenInterpret
This required pulling in a bunch of function definitions.

This also required me to fix previously defined Wren*Fn definitions to be wrapped in Option<> to cover the null ptr case.

I used bindgen -- see (unused) example_wren_bindings.rs -- to pull these from, but then I hand edited in some places.

Eventually would be nice to use bindgen for some of this, but it seems more designed for generating rust from a .h for you to call out to C, rather than generating rust from a .h to let you implement the .h api in rust.