Add wrenGetSlotHandle, wrenReleaseHandle, wrenSetSlotHandle
No clue if these work, but we'll know when this finally links and runs.  This resolves 3 more link errors.  Only 7 more to go!

For now WrenHandle is just a Box<Value>.  It's possible we could just type alias WrenHandle to Box<Value> and it would just work. Something to try once everything links.

It's also possible we'll want a fancier WrenHandle which can help with debugging Handle leaks or other errors in API usage.