Make benchmarks 1.3 to 2.3x better by removing malloc.
We've known for a while that splitting out the args to pass to functions (especially primitives) causes a needless malloc.  It showed up as ~20% of the profile for the binary_trees benchmark, but as these things go, when removing it we actually got a 30% gain (you gain some on the free size, plus also sometimes from the locks and other hidden costs, plus also sometimes from improving cache locality).

This win showed up even more pronounced in other benchmarks (e.g. string_equals or for) even though I didn't profile those.

I'm not sure that this is even as efficient as we could get.  If we passed the Drain iterator to primitives we could probably avoid some clones.  Those aren't top of the benchmark, but I do think they show up.

Anyway, taking the win and will look further in other patches.

I also cleaned up / made more consistent access to named args like "key" in a few places.  Similarly in order to do this optimization, I moved the frame.push() out of call_method into the caller.  It was part of an earlier attempt and I don't think it was probably needed in the final, but it seemed clean enough to re-use FunctionNext::Return instead of "None" so I kept it.

I also deployed the Symbol type in a few more places.