Make Arity its own type
I'm not sure this is actually a good change.  But it does add some type-safety to Arity.  At the cost of having to implement a somewhat awkward Arity type and then also be uncertain where to use it.

On (concerning) thing this did reveal is that we seem to sometimes use Arity with and without implicit this.  Unclear when referring to an Arity type if it includes the implicit +1 for "this" for methods.  That seems wrong.

Since all the tests still pass, I'm going to take this as a small improvement and iterate later from here.

Also fixed Constant Debug to not say Constant(0) but rather just "0".