Add VM.interpret_bytes for C-API
Added a VM.interpret_bytes.  The C API expects to take a CString of (possibly invalid) utf8.  Since I had to add that for the C-API, I also wired it up for the Rust API and made it report decoding failures similarly to how wren_c would.  It's not exactly identical, since we don't even attempt to parse if the string doesn't decode as valid utf8, but it was enough to pass 12 of the "regression/442*" (really utf8 decoding error) tests.

Without this change, the CAPI simply panicks (due to unwrapping None due to a failed decode) whenever passed invalid utf8.  This seems at least slightly better than panicking.