Make InputManager an Iterator
This is only partial.  We're not actually using any of the iterator methods yet.

I tried to use Iterator::skip_while to replace our existing skip_while, but can't, since it returns an iterator which we're supposed to use to replace our existing one.

I did however update the types to use references (somewhat unnecessary) in an effort to match Iterator patterns.