Use an explicit Compiler stack instead of linked list
I should have done this long ago.  But untangling this ownership to make it more rust-friendly, blocks at least 3 changes.  Attempting to work with two compilers at the same time (independently mutable) was difficult/impossible with the previous design in rust.

I did not attempt to refactor much here, just change the ownership model.  We likely can move to a "grab me the enclosing class compiler" module instead of the current callback effects.  We can also add a parent_compiler, etc. method to grab earlier compilers as well.  Future changes.

I was amused to learn that "parent" was only used inside the Compiler class, so removing it turned out to be very easy.  It's only passed to the Compiler constructor to handle inheriting scope depth (which also may not be needed).