Add Compiler::enclosing_class_compiler and refactor around it
It's not clear if this is better or not.  I guess it's good in the sense of having the .rev search in one place instead of inviting bugs of forgetting the .rev() one time.

This matches more closely what wren_c does (now that such is possible in wren_rust).  I suspect enclosing_class_compiler will end up used in other code.