Add Partial closures support
This still crashes, but it also passes THIRTY ONE more tests, so committing as a checkpoint.

Fixed non-local closure creation to start at +2 rather than +1.  The value being closed over is in compiler +0, the direct child is +1 and all "non-local" closures start at +2.  Just an error when converting wren_c's recursive algorithm to my looping one.

I was confused about how closeUpvalues worked in wren_c.  It is infact needed to be issued for each local being closed, as it also issues the "pop" for that local.  Fixed.  This allowed us to delete the PopsResult helper struct as we're no longer trying to coalesce multiple closeUpvalues into one.

Added StackOffset (originally called StackLocation, but also kept a fiber, this was a partial refactor).  Upvalues now have a storage enum, mostly to allow someone to hold onto the upvalue and later swap its internal storage type.

I discovered through this that some of the complexity with upvalue opcodes is due to wren treating "return" sometimes like "yield".  So in the case of "return", the compiler never issues pops (or upvalue closures) like it does in the normal going out of scope case.

Much of the upvalue logic has to be hoisted up out of run_frame_in_fiber since run_frame_in_fiber cannot modify the fiber.